Renamed config files
