+++ config/webpack/webpack.config.common.js
