+++ config/webpack/webpack.config.prod.js
@@ -1,5 +1,5 @@
 const webpack = require('webpack');
-const webpackConfig = require('./webpack.common');
+const webpackConfig = require('./webpack.config.common');
 const helpers = require('../helpers');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const { prodLoaders } = require('./loaders');
