+++ webpack.config.js
@@ -1,1 +1,1 @@
-module.exports = require('./config/webpack/webpack.dev');
+module.exports = require('./config/webpack/webpack.config.dev');
